requirements.txt
flask
supabase-py
python-dotenv
